requirements.txt
# === Test Tools ===
pytest==8.4.0
playwright==1.52.0
pytest-playwright==0.7.0

# === HTTP Support ===
requests==2.32.4

# === Code Quality Tools ===
flake8==7.2.0
black==25.1.0

# === Allure Tools ===
allure-pytest==2.14.3
allure-python-commons==2.14.3
attrs==25.3.0